+++ lib/storage/metadata/in_memory/ListMultipartUploadsResult.js
@@ -4,22 +4,26 @@ export class ListMultipartUploadsResult extends ListResult {
     constructor() {
         super();
         this.Uploads = [];
+        this.NextKeyMarker = undefined;
+        this.NextUploadIdMarker = undefined;
     }
 
     addUpload(uploadInfo) {
         this.Uploads.push({
-            Key: decodeURIComponent(uploadInfo.key),
-            UploadId: uploadInfo.uploadId,
-            Initiator: {
-                ID: uploadInfo.initiatorID,
-                DisplayName: uploadInfo.initiatorDisplayName,
+            key: decodeURIComponent(uploadInfo.key),
+            value: {
+                UploadId: uploadInfo.uploadId,
+                Initiator: {
+                    ID: uploadInfo.initiatorID,
+                    DisplayName: uploadInfo.initiatorDisplayName,
+                },
+                Owner: {
+                    ID: uploadInfo.ownerID,
+                    DisplayName: uploadInfo.ownerDisplayName,
+                },
+                StorageClass: uploadInfo.storageClass,
+                Initiated: uploadInfo.initiated,
             },
-            Owner: {
-                ID: uploadInfo.ownerID,
-                DisplayName: uploadInfo.ownerDisplayName,
-            },
-            StorageClass: uploadInfo.storageClass,
-            Initiated: uploadInfo.initiated,
         });
         this.MaxKeys += 1;
     }
