+++ lib/storage/metadata/in_memory/backend.js
@@ -3,6 +3,7 @@ import naturalCompare from 'natural-compare-lite';
 
 import { markerFilter, prefixFilter } from './bucket_utilities';
 import { ListBucketResult } from './ListBucketResult';
+import getMultipartUploadListing from './getMultipartUploadListing';
 import metadata from './metadata';
 
 const defaultMaxKeys = 1000;
@@ -173,6 +174,20 @@ const metastore = {
                 }
                 return cb(null, response);
             });
+    },
+
+    listMultipartUploads(bucketName, listingParams, log, cb) {
+        metastore.getBucketAttributes(bucketName, log, (err, bucket) => {
+            if (bucket === undefined) {
+                // no on going multipart uploads, return empty listing
+                return cb(null, {
+                    IsTruncated: false,
+                    NextMarker: undefined,
+                    MaxKeys: 0,
+                });
+            }
+            getMultipartUploadListing(bucket, listingParams, cb);
+        });
     },
 };
