+++ lib/storage/metadata/in_memory/bucket_utilities.js
@@ -57,34 +57,6 @@ export function prefixFilter(prefix, array) {
     return array;
 }
 
-export function findNextMarker(currentIndex, array, responseObject) {
-    let delimiterIndex;
-    let delimitedKey;
-    // Iterate through remainder of array to
-    // find next key that would NOT get rolled
-    // up into an existing CommonPrefix
-    for (let i = currentIndex; i < array.length; i++) {
-        // If this function is called for a bucket listing, the key
-        // will be array[i].
-        // If this is for a multipart
-        // upload listing, the key will be array[i].key.
-        const key = array[i].key ? array[i].key : array[i];
-        if (responseObject.Delimiter) {
-            delimiterIndex = key.indexOf(responseObject.Delimiter);
-            delimitedKey = key.slice(0, delimiterIndex + 1);
-        }
-        if (!responseObject.hasCommonPrefix(delimitedKey)) {
-            // If this function is called for a bucket listing,
-            // just return the key.  If this is for a multipart
-            // upload listing, return an array containing the
-            // key and the uploadId.
-            const result = array[i].key ? [key, array[i].uploadId] : key;
-            return result;
-        }
-    }
-    return null;
-}
-
 export function isKeyInContents(responseObject, key) {
     return responseObject.Contents.some(val => val.key === key);
 }
